refactor get_url functions to re-useable layer file
